Changed coin score to 100 for balance
This is a game balance change that is intended to change the focus of how it is played and how the maps are made.  The coins will be bonuses if you are willing to do the platforming to get them.  you can complete the game without getting many coins and see the whole game and that is an accomplishment but the more you explore the higher score you can get.